Fixed Turret Rotation and updated soda mine
