+++ My project/Assets/Alexander/Traps_Scripts/Advanced_Proyectile.cs.meta
