+++ My project/Assets/Alexander/Traps_Scripts/Lightning_Trap.cs
