+++ My project/Assets/Alexander/Traps_Scripts/Soda_Mine.cs.meta
